server/src/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
greenlet==3.0.1
Mako==1.3.0
MarkupSafe==2.1.3
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
asyncpg==0.27.0